requirements.txt
# =============================================================================
# NetSentinel Requirements - Consolidated Dependencies
# =============================================================================
# This file contains all dependencies for NetSentinel in a consolidated format.
# Core dependencies are required for basic functionality.
# Optional dependencies enable additional features but are not required.
# Development dependencies are for development, testing, and CI/CD.
# =============================================================================

# =============================================================================
# CORE DEPENDENCIES (Required for basic NetSentinel functionality)
# =============================================================================

# Core Framework & Networking
Twisted>=24.11.0
zope.interface>=7.2
requests>=2.31.0
urllib3>=2.0.7
simplejson>=3.16.0

# Security & Cryptography
cryptography>=41.0.0  # Updated from 38.0.1 for security
pyasn1>=0.4.5
pyOpenSSL>=22.1.0
service-identity>=21.1.0
bcrypt>=4.0.0  # Updated from 3.1.7 for security

# Authentication & Authorization
passlib>=1.7.1

# Data Processing & Serialization
Jinja2>=3.0.1
PyPDF2>=1.26.0
fpdf>=1.7.2

# Protocol Libraries
ntlmlib>=0.72
hpfeeds>=3.0.0

# Web Framework (Core)
flask>=2.3.0

# Message Queue & Caching
kafka-python>=2.0.2
redis>=4.6.0

# Networking & Packet Analysis
scapy>=2.5.0
pcapy-ng>=1.0.0

# =============================================================================
# OPTIONAL DEPENDENCIES (Enable additional features)
# =============================================================================

# API Server (REST API functionality)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Machine Learning & Anomaly Detection
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
lightning>=2.0.0  # Updated from pytorch-lightning
anomalib>=1.1.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=10.0.0
scikit-learn>=1.0.0

# Enterprise Database Integration
elasticsearch>=8.0.0
influxdb-client>=1.36.0

# Monitoring & Observability
prometheus-client>=0.17.0
psutil>=5.9.0

# Enhanced Networking & Async I/O
aiohttp>=3.8.0

# Data Models & Validation
pydantic>=2.0.0
dataclasses-json>=0.6.0

# MQTT Support
asyncio-mqtt>=0.13.0

# JWT Authentication
PyJWT>=2.8.0

# =============================================================================
# DEVELOPMENT & TESTING DEPENDENCIES (For development, testing, CI/CD)
# =============================================================================

# Testing Framework
pytest>=7.4.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-html>=3.1.0
pytest-playwright>=0.3.0

# Playwright for Web Testing
playwright>=1.30.0

# Test Data & Fixtures
faker>=13.0.0
factory-boy>=3.2.0

# Code Quality & Linting
black>=22.0.0
flake8>=4.0.0
flake8-typing-imports>=1.12.0
isort>=5.10.0
mypy>=0.950
pylint>=2.12.0
bandit>=1.7.0
safety>=2.2.0

# Code Coverage
coverage>=6.3.0
coverage-badge>=1.1.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0

# Development Tools
pre-commit>=2.17.0
tox>=3.24.0
pip-tools>=6.6.0

# Performance Testing
locust>=2.15.0,<2.30.0

# Container Testing
docker>=5.0.0
docker-compose>=1.29.0

# Kubernetes Testing
kubernetes>=21.7.0

# API Testing
requests-mock>=1.9.0
responses>=0.20.0
httpx>=0.23.0

# Database Testing
testcontainers>=3.7.0

# Security Testing
sqlmap>=1.6.0

# Web Framework Testing
flask-testing>=0.8.1

# Environment Management
python-dotenv>=0.19.0

# Type Checking
typing-extensions>=4.0.0

# Logging Testing
structlog>=21.5.0

# Time Testing
freezegun>=1.1.0

# API Documentation Testing
drf-spectacular>=0.24.0
openapi-spec-validator>=0.4.0

# Additional Development Dependencies
asyncio-mqtt>=0.13.0
PyJWT>=2.8.0